+++ usecase/src/main/kotlin/dev/agner/portfolio/usecase/index/model/TheirIndexValue.kt
@@ -1,10 +1,11 @@
 package dev.agner.portfolio.usecase.index.model
 
 import kotlinx.datetime.LocalDate
+import java.math.BigDecimal
 
 data class TheirIndexValue(
     val date: LocalDate,
-    val value: Double,
+    val value: BigDecimal,
 ) {
     fun toCreation() = IndexValueCreation(
         date = date,
